Use a single global infinite promise (#34)
It's not super clear why the code is written this way, but I assume it's due to not being familiar with the pattern. If you want to wait until the `freeze` prop changes, it's enough to throw a Promise that never resolves.

This fixes an issue that would occur on initial mount with `createRoot` and/or Fabric. The issue is that React throws away uncommitted trees if they suspend during mount, and this includes throwing away `useRef`. So stashing Promises in `useRef` is wrong. Having one at the top level is both safe and simpler.